+++ docs/html/search/all_3.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['findandruncommand_11',['findAndRunCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a97631715e21c262544a228377688d593',1,'ace_utils::cli::CommandDispatcher']]],
-  ['findcommand_12',['findCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ad59ab4795f6834ccf252dce8b0682c52',1,'ace_utils::cli::CommandDispatcher']]]
+  ['espeepromadapter_14',['EspEepromAdapter',['../classace__utils_1_1crc__eeprom_1_1EspEepromAdapter.html',1,'ace_utils::crc_eeprom::EspEepromAdapter&lt; E &gt;'],['../classace__utils_1_1crc__eeprom_1_1EspEepromAdapter.html#a34548c1e2db594d905d7b564fdd6910a',1,'ace_utils::crc_eeprom::EspEepromAdapter::EspEepromAdapter()']]]
 ];