+++ docs/html/search/all_a.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['parentgroup_28',['parentGroup',['../structace__utils_1_1mode__group_1_1ModeGroup.html#a8e8e2d3b47f11a516d70d2664db9bcde',1,'ace_utils::mode_group::ModeGroup']]],
-  ['printhelp_29',['printHelp',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ac9a7dea42f2dc74729e524e71c5b9d54',1,'ace_utils::cli::CommandDispatcher']]],
-  ['printlineerror_30',['printLineError',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a0db9822394faea14916353985c9ff7b0',1,'ace_utils::cli::CommandDispatcher']]]
+  ['parentgroup_32',['parentGroup',['../structace__utils_1_1mode__group_1_1ModeGroup.html#a8e8e2d3b47f11a516d70d2664db9bcde',1,'ace_utils::mode_group::ModeGroup']]],
+  ['printhelp_33',['printHelp',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ac9a7dea42f2dc74729e524e71c5b9d54',1,'ace_utils::cli::CommandDispatcher']]],
+  ['printlineerror_34',['printLineError',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a0db9822394faea14916353985c9ff7b0',1,'ace_utils::cli::CommandDispatcher']]]
 ];