+++ docs/html/search/classes_2.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['modegroup_46',['ModeGroup',['../structace__utils_1_1mode__group_1_1ModeGroup.html',1,'ace_utils::mode_group']]],
-  ['modenavigator_47',['ModeNavigator',['../classace__utils_1_1mode__group_1_1ModeNavigator.html',1,'ace_utils::mode_group']]]
+  ['espeepromadapter_57',['EspEepromAdapter',['../classace__utils_1_1crc__eeprom_1_1EspEepromAdapter.html',1,'ace_utils::crc_eeprom']]]
 ];