+++ docs/html/search/functions_1.js
@@ -1,9 +1,4 @@
 var searchData=
 [
-  ['changegroup_50',['changeGroup',['../classace__utils_1_1mode__group_1_1ModeNavigator.html#adaae7dc240a20a8dc623a5168fee4981',1,'ace_utils::mode_group::ModeNavigator']]],
-  ['changemode_51',['changeMode',['../classace__utils_1_1mode__group_1_1ModeNavigator.html#a20914625815dbedaa60d12d407e7bf89',1,'ace_utils::mode_group::ModeNavigator']]],
-  ['commanddispatcher_52',['CommandDispatcher',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ad1ab7be3d70937bf978c6ab2d2ceaf9b',1,'ace_utils::cli::CommandDispatcher']]],
-  ['commandhandler_53',['CommandHandler',['../classace__utils_1_1cli_1_1CommandHandler.html#a50a4af8c63b7bcafe0ed7e8269d3da50',1,'ace_utils::cli::CommandHandler::CommandHandler(const char *name, const char *helpString)'],['../classace__utils_1_1cli_1_1CommandHandler.html#a846e22124c7239103bc75bc0baf0e598',1,'ace_utils::cli::CommandHandler::CommandHandler(const __FlashStringHelper *name, const __FlashStringHelper *helpString)']]],
-  ['commandmanager_54',['CommandManager',['../classace__utils_1_1cli_1_1CommandManager.html#a367a2cbd734cac13b875a7fa44224ff5',1,'ace_utils::cli::CommandManager']]],
-  ['crceeprom_55',['CrcEeprom',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#ad6f0e26adae30c2397313dc1547d0cfb',1,'ace_utils::crc_eeprom::CrcEeprom']]]
+  ['begin_64',['begin',['../classace__utils_1_1crc__eeprom_1_1IEepromAdapter.html#a26622e062b7b3d49bc479d64633acd46',1,'ace_utils::crc_eeprom::IEepromAdapter::begin()'],['../classace__utils_1_1crc__eeprom_1_1AvrEepromAdapter.html#ab63f1275794c652ff1cf225a465b0989',1,'ace_utils::crc_eeprom::AvrEepromAdapter::begin()'],['../classace__utils_1_1crc__eeprom_1_1EspEepromAdapter.html#ab50ffbfd2dee315be70a42d84baf2a91',1,'ace_utils::crc_eeprom::EspEepromAdapter::begin()'],['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#afa98dd019e95fad991a77296eec0e728',1,'ace_utils::crc_eeprom::CrcEeprom::begin()']]]
 ];