+++ docs/html/search/functions_2.js
@@ -1,5 +1,10 @@
 var searchData=
 [
-  ['findandruncommand_56',['findAndRunCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a97631715e21c262544a228377688d593',1,'ace_utils::cli::CommandDispatcher']]],
-  ['findcommand_57',['findCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ad59ab4795f6834ccf252dce8b0682c52',1,'ace_utils::cli::CommandDispatcher']]]
+  ['changegroup_65',['changeGroup',['../classace__utils_1_1mode__group_1_1ModeNavigator.html#adaae7dc240a20a8dc623a5168fee4981',1,'ace_utils::mode_group::ModeNavigator']]],
+  ['changemode_66',['changeMode',['../classace__utils_1_1mode__group_1_1ModeNavigator.html#a20914625815dbedaa60d12d407e7bf89',1,'ace_utils::mode_group::ModeNavigator']]],
+  ['commanddispatcher_67',['CommandDispatcher',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ad1ab7be3d70937bf978c6ab2d2ceaf9b',1,'ace_utils::cli::CommandDispatcher']]],
+  ['commandhandler_68',['CommandHandler',['../classace__utils_1_1cli_1_1CommandHandler.html#a50a4af8c63b7bcafe0ed7e8269d3da50',1,'ace_utils::cli::CommandHandler::CommandHandler(const char *name, const char *helpString)'],['../classace__utils_1_1cli_1_1CommandHandler.html#a846e22124c7239103bc75bc0baf0e598',1,'ace_utils::cli::CommandHandler::CommandHandler(const __FlashStringHelper *name, const __FlashStringHelper *helpString)']]],
+  ['commandmanager_69',['CommandManager',['../classace__utils_1_1cli_1_1CommandManager.html#a367a2cbd734cac13b875a7fa44224ff5',1,'ace_utils::cli::CommandManager']]],
+  ['commit_70',['commit',['../classace__utils_1_1crc__eeprom_1_1IEepromAdapter.html#a7e0a7eb29299bbeb9d5444f1865f81de',1,'ace_utils::crc_eeprom::IEepromAdapter::commit()'],['../classace__utils_1_1crc__eeprom_1_1AvrEepromAdapter.html#affad9de69e6dc55616c3f30bbd17c7ba',1,'ace_utils::crc_eeprom::AvrEepromAdapter::commit()'],['../classace__utils_1_1crc__eeprom_1_1EspEepromAdapter.html#a9cb0ee4c0189f77c5182f8cc2f6131e0',1,'ace_utils::crc_eeprom::EspEepromAdapter::commit()']]],
+  ['crceeprom_71',['CrcEeprom',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#aa7f7bbf0f1eeff6430792a4612f0fc0e',1,'ace_utils::crc_eeprom::CrcEeprom']]]
 ];