+++ docs/html/search/functions_7.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['printhelp_67',['printHelp',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ac9a7dea42f2dc74729e524e71c5b9d54',1,'ace_utils::cli::CommandDispatcher']]],
-  ['printlineerror_68',['printLineError',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a0db9822394faea14916353985c9ff7b0',1,'ace_utils::cli::CommandDispatcher']]]
+  ['isargequal_81',['isArgEqual',['../classace__utils_1_1cli_1_1CommandHandler.html#ab664de864c17a35e0bbb6c75176684cb',1,'ace_utils::cli::CommandHandler::isArgEqual(const char *arg, const char *token)'],['../classace__utils_1_1cli_1_1CommandHandler.html#aecf6a79d06930a7d91c8862274af698a',1,'ace_utils::cli::CommandHandler::isArgEqual(const char *arg, const __FlashStringHelper *token)']]]
 ];