+++ docs/html/search/functions_a.js
@@ -1,4 +1,9 @@
 var searchData=
 [
-  ['tokenize_76',['tokenize',['../classace__utils_1_1cli_1_1CommandDispatcher.html#a535d819984ddd20524a00837d14bfa1a',1,'ace_utils::cli::CommandDispatcher']]]
+  ['read_86',['read',['../classace__utils_1_1crc__eeprom_1_1IEepromAdapter.html#a42c68c82eeea852da55bad99e0af6f4a',1,'ace_utils::crc_eeprom::IEepromAdapter::read()'],['../classace__utils_1_1crc__eeprom_1_1AvrEepromAdapter.html#ae386c6508f316a1fb45ab9176f5d8aba',1,'ace_utils::crc_eeprom::AvrEepromAdapter::read()'],['../classace__utils_1_1crc__eeprom_1_1EspEepromAdapter.html#ac9a01cd19e235f14f9d21bb9facf0a64',1,'ace_utils::crc_eeprom::EspEepromAdapter::read()']]],
+  ['readdatawithcrc_87',['readDataWithCrc',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#a5e45d0ea63efe500a0aa240f8d6a3b06',1,'ace_utils::crc_eeprom::CrcEeprom']]],
+  ['readwithcrc_88',['readWithCrc',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#aefa1a740fbada5780ea9d99eb7a74ec9',1,'ace_utils::crc_eeprom::CrcEeprom']]],
+  ['run_89',['run',['../classace__utils_1_1cli_1_1CommandHandler.html#a82e846fdd080bfb36b13a1f903cb7def',1,'ace_utils::cli::CommandHandler']]],
+  ['runcommand_90',['runCommand',['../classace__utils_1_1cli_1_1CommandDispatcher.html#ae20ae336f3740b1faceecdf0babfa55e',1,'ace_utils::cli::CommandDispatcher']]],
+  ['runcoroutine_91',['runCoroutine',['../classace__utils_1_1cli_1_1CommandManager.html#aee686417c164a6d1a243f1d53ab7aa6d',1,'ace_utils::cli::CommandManager::runCoroutine()'],['../classace__utils_1_1cli_1_1StreamLineReader.html#a6171d73022c91c7f826d1aaf42357ced',1,'ace_utils::cli::StreamLineReader::runCoroutine()']]]
 ];