+++ docs/html/search/pages_1.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['command_20line_20interface_20_28cli_29_86',['Command Line Interface (CLI)',['../md__home_brian_src_AceUtils_src_cli_README.html',1,'']]]
+  ['command_20line_20interface_20_28cli_29_109',['Command Line Interface (CLI)',['../md__home_brian_src_AceUtils_src_cli_README.html',1,'']]],
+  ['crc_20eeprom_110',['CRC EEPROM',['../md__home_brian_src_AceUtils_src_crc_eeprom_README.html',1,'']]]
 ];