+++ docs/html/search/variables_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['kcrcsize_80',['kCrcSize',['../classace__utils_1_1crc__eeprom_1_1CrcEeprom.html#a2f2fcf3066d500a77cfbbf1e03437a8e',1,'ace_utils::crc_eeprom::CrcEeprom']]]
+  ['modes_104',['modes',['../structace__utils_1_1mode__group_1_1ModeGroup.html#a1957bb0d8c4cc3fa53fa12a078d45558',1,'ace_utils::mode_group::ModeGroup']]]
 ];